protecting user notes with middleware

+++ backend/src/controllers/notes.ts
@@ -1,22 +1,29 @@
 import { NextFunction, Request, RequestHandler, Response } from "express"
 
 import NoteModel from '../models/note'
+import { assertIsDefined } from "../util/assertIsDefined"
 import createHTTPError from 'http-errors'
 import mongoose from 'mongoose'
 
 export const getNotes: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
+    const authenticatedUserId = req.session.userId
     try {
-        const notes = await NoteModel.find({}).exec()
+        assertIsDefined(authenticatedUserId)
+
+        const notes = await NoteModel.find({ userId: authenticatedUserId }).exec()
+
         return res.status(200).json(notes)
     } catch (err) {
         next(err)
     }
 }
 
-
 export const getNote: RequestHandler = async(req, res, next) => {
+    const authenticatedUserId = req.session.userId
     const noteId = req.params.id
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHTTPError(400, "Invalid note id")
         }
@@ -25,6 +32,10 @@ export const getNote: RequestHandler = async(req, res, next) => {
 
         if (!note) {
             throw createHTTPError(404, "Note not found")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHTTPError(401, 'You cannot access this note.')
         }
 
         return res.status(200).json(note)
@@ -39,17 +50,23 @@ interface CreateNodeBody {
 }
 
 export const createNote: RequestHandler<unknown, unknown, CreateNodeBody, unknown> = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     const title = req.body.title
     const text = req.body.text
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if (!title) {
             throw createHTTPError(400, "Note must have a title")
         }
+
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title,
             text
         })
+
         res.status(201).json(newNote)
     } catch (err) {
         next(err)
@@ -65,11 +82,14 @@ interface UpdateNodeBody {
 }
 
 export const patchNote: RequestHandler<UpdateNoteParams, undefined, UpdateNodeBody, undefined> = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     const noteId = req.params.id
     const newTitle = req.body.title
     const newText = req.body.text
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHTTPError(400, "Invalid note id")
         }
@@ -84,6 +104,10 @@ export const patchNote: RequestHandler<UpdateNoteParams, undefined, UpdateNodeBo
             throw createHTTPError(404, "Note not found")
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHTTPError(401, 'You cannot access this note.')
+        }
+
         note.title = newTitle;
         note.text = newText;
 
@@ -96,8 +120,12 @@ export const patchNote: RequestHandler<UpdateNoteParams, undefined, UpdateNodeBo
 }
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     const noteId = req.params.id
+
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHTTPError(400, "Invalid note id")
         }
@@ -106,6 +134,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHTTPError(404, "Note not found")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHTTPError(401, 'You cannot access this note.')
         }
 
         await note.deleteOne()
